Move meta models into seperate files
